refactor: replace reloader with loader package for configuration management and streamline initialization process once again
